+++ examples/03_spam_filter.rs
@@ -14,6 +14,7 @@ use nostr_relay_builder::{
 };
 use nostr_sdk::prelude::*;
 use std::net::SocketAddr;
+use std::sync::Arc;
 
 /// Custom processor that filters spam
 #[derive(Debug, Clone)]
@@ -26,7 +27,7 @@ impl EventProcessor for SpamFilterProcessor {
     async fn handle_event(
         &self,
         event: Event,
-        _custom_state: &mut (),
+        _custom_state: Arc<tokio::sync::RwLock<()>>,
         context: EventContext<'_>,
     ) -> RelayResult<Vec<StoreCommand>> {
         let content_lower = event.content.to_lowercase();