+++ examples/08_multi_tenant.rs
@@ -14,6 +14,7 @@ use nostr_relay_builder::{
 };
 use nostr_sdk::prelude::*;
 use std::net::SocketAddr;
+use std::sync::Arc;
 
 /// Simple processor that logs which subdomain received the event
 #[derive(Debug, Clone)]
@@ -24,7 +25,7 @@ impl EventProcessor for MultiTenantProcessor {
     async fn handle_event(
         &self,
         event: Event,
-        _custom_state: &mut (),
+        _custom_state: Arc<tokio::sync::RwLock<()>>,
         context: EventContext<'_>,
     ) -> RelayResult<Vec<StoreCommand>> {
         // Log subdomain info